% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{batch_correct_df}
\alias{batch_correct_df}
\title{Run batch correction on data}
\usage{
batch_correct_df(
  df,
  label = NULL,
  xdim = 8,
  ydim = 8,
  rlen = 10,
  mode = c("online", "batch", "pbatch"),
  parametric = TRUE,
  method = c("ComBat", "ComBat_seq"),
  cluster_method = c("kohonen", "flowsom", "fusesom", "kmeans"),
  distf = c("euclidean", "sumofsquares", "cosine", "manhattan", "chebyshev"),
  nClus = NULL,
  mc.cores = 1,
  pb = FALSE,
  ref.batch = NULL,
  seed = 473,
  covar = NULL,
  anchor = NULL,
  markers = NULL,
  norm_method = c("scale", "rank", "CLR_seu", "CLR_med", "CLR", "qnorm", "none"),
  ties.method = "average",
  ...
)
}
\arguments{
\item{df}{tibble or seurat with expression values}

\item{label}{The cluster or cell type label. Either as a column name or vector.}

\item{xdim}{The x-dimension size of the SOM. If NULL, gridsize is computed using \code{FuseSOM::computeGridSize}}

\item{ydim}{The y-dimension size of the SOM. If NULL, gridsize is computed using \code{FuseSOM::computeGridSize}}

\item{rlen}{Number of times the data is presented to the SOM network}

\item{mode}{type of learning algorithm.}

\item{parametric}{Default: TRUE. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used.}

\item{method}{Default: "ComBat". Choose "ComBat" for cytometry data and "ComBat_seq" for bulk RNAseq data.}

\item{cluster_method}{Cluster method to use. Defaults to kohonen. Options: c("kohonen", "flowsom", "fusesom", "kmeans")}

\item{distf}{Distance metric used for kohonen ("sumofsquares", "euclidean", "manhattan") or FlowSOM ("euclidean", "cosine", "manhattan", "chebyshev")}

\item{nClus}{(Usable with FlowSOM, FuseSOM, and kmeans) Number of clusters to export}

\item{mc.cores}{Number of cores for parallelization}

\item{pb}{Progress bar for parallelization}

\item{ref.batch}{Optional. A string of the batch that should be used as the reference for batch adjustments.}

\item{seed}{The seed to use when creating the SOM.}

\item{covar}{The covariate ComBat should use. Can be a vector or a column name in the input tibble.
If NULL, no covar will be used}

\item{anchor}{Experimental: A column or vector specifying which samples are replicates and which are not. If specified, this column will be used as a covariate in ComBat. Be aware that it may be confounded with the condition.}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{norm_method}{Normalization method. Should be either 'rank', 'scale', 'CLR', 'CLR_seu', 'CLR_med', or 'qnorm'. Default: 'scale'. CLR is recommended for ADT normalization. CLR_seu mimicks the Seurat implementation of CLR. CLR_med uses the median instead of geometric mean.}

\item{ties.method}{The method to handle ties, when using rank. Default: 'average'. See ?rank for other options.}

\item{...}{Arguments passed to \code{create_som_sce}}
}
\description{
This is a wrapper function for the cyCombine batch correction workflow.
To run the workflow manually, type "batch_correct" to see the source code of this wrapper and follow along or read the vignettes on the GitHub page \url{https://github.com/biosurf/cyCombine}.
}
\examples{
\dontrun{
corrected <- uncorrected \%>\%
  batch_correct(markers = markers,
  covar = "condition")
  }
}
\seealso{
Other batch: 
\code{\link{batch_correct_seurat}()},
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{normalize}()},
\code{\link{quantile_norm}()}
}
\concept{batch}
