% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{compile_fcs}
\alias{compile_fcs}
\title{Compile all .fcs files in a directory to a flowset}
\usage{
compile_fcs(
  data_dir,
  meta_filename = NULL,
  sample_col = NULL,
  batch_col = "Batch",
  filename_col = "FCS_name",
  pattern = "\\\\.fcs"
)
}
\arguments{
\item{data_dir}{Directory containing the .fcs files}

\item{meta_filename}{Filename of the metadata file. If NULL, sample and batch ids are not predicted nor returned}

\item{sample_col}{The column in the metadata filename containing the sample ids. If NULL, sample ids will be the file names}

\item{batch_col}{The column in the metadata filename containing the sample ids}

\item{pattern}{The pattern to use to find the files in the folder.}
}
\description{
Compile all .fcs files in a directory to a flowset
}
