% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{compute_emd}
\alias{compute_emd}
\title{Compute EMD}
\usage{
compute_emd(
  df,
  binSize = 0.1,
  cell_col = "label",
  batch_col = "batch",
  markers = NULL
)
}
\arguments{
\item{df}{Dataframe to compute the EMD of}

\item{binSize}{The size of bins to use when binning data}

\item{cell_col}{Column name of df that contains cell population labels (or clusters)}

\item{batch_col}{Column name of df that contains batch numbers}

\item{markers}{Vector of the markers to calculate EMD for. If NULL, \code{\link{get_markers}} will be used to find markers}
}
\description{
Given a dataframe, the Earth Movers Distance is computed, using the emdist package.
A comparison matrix is returned containing the maximum distance for each marker for each group of cells.
}
