% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lisi.R
\name{compute_lisi}
\alias{compute_lisi}
\title{Compute Local Inverse Simpson's Index (LISI)}
\usage{
compute_lisi(
  df,
  markers = get_markers(df),
  label_cols = "batch",
  perplexity = 30,
  nn_eps = 0
)
}
\arguments{
\item{df}{A data.frame with cells (rows) and markers in columns.}

\item{markers}{Character vector or integer indices of columns containing features.
Default is all numeric columns excluding metadata specified in \code{label_cols}.}

\item{label_cols}{Character vector of column names in \code{df} to compute LISI for.}

\item{perplexity}{The effective number of neighbors.}

\item{nn_eps}{Error bound for nearest neighbor search.}
}
\value{
A tibble with the original metadata plus new columns \verb{lisi_<label>} for each label.
}
\description{
NOTE: reimplemented from https://github.com/immunogenomics/LISI
}
