% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{compute_mad}
\alias{compute_mad}
\title{Compute MAD}
\usage{
compute_mad(df, batch_col = "batch", markers = NULL)
}
\arguments{
\item{df}{Dataframe to compute the MADs of}

\item{batch_col}{Column name of df that contains batch numbers}

\item{markers}{Vector of the markers to calculate EMD for. If NULL, \code{\link{get_markers}} will be used to find markers}
}
\description{
Given a dataframe, the Median Absolute Deviation (MAD) is calculated per-marker, per-batch
}
\seealso{
Other mad: 
\code{\link{evaluate_mad}()}
}
\concept{mad}
