% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{convert_flowset}
\alias{convert_flowset}
\title{Convert a flowset into a workable dataframe}
\usage{
convert_flowset(
  flowset,
  metadata = NULL,
  filename_col = "filename",
  sample_ids = NULL,
  batch_ids = NULL,
  condition = NULL,
  down_sample = TRUE,
  sample_size = 5e+05,
  seed = 473,
  panel = NULL,
  panel_channel = "fcs_colname",
  panel_antigen = "antigen"
)
}
\arguments{
\item{flowset}{The flowset to convert}

\item{metadata}{Optional: Can be either a filename or data.frame of the metadata file. Please give the full path from working directory to metadata file}

\item{filename_col}{Optional: The column in the metadata containing the fcs filenames. Needed if metadata is given, but sample_ids is not}

\item{sample_ids}{Optional: If a character, it should be the sample column in the metadata. If its a vector, it should have the same length as the total flowset. If NULL, sample ids will be the file names}

\item{batch_ids}{Optional: If a character, it should be the column in the metadata containing the batch ids. If its a vector, it should have the same length as the total flowset.}

\item{condition}{Optional: The column in the metadata containing the condition. Will be used as the covariate in ComBat, but can be specified later.}

\item{down_sample}{If TRUE, the output will be down-sampled to size sample_size}

\item{sample_size}{The size to down-sample to}

\item{seed}{The seed to use for down-sampling}

\item{panel}{Optional: Panel as a fileanme or data.frame. Is used to define colnames from the panel_antigen column}

\item{panel_channel}{Optional: Only used if panel is given. It is the column name in the panel data.frame that contains the channel names}

\item{panel_antigen}{Optional: Only used if panel is given. It is the column name in the panel data.frame that contains the antigen names}
}
\description{
Convert a flowset into a workable dataframe
}
\examples{
df <- convert_flowset(flowset = fcs$fcs_raw, sample_ids = fcs$sample_ids, batch_ids = fcs$batch_ids, down_sample = FALSE)
}
