% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{create_sample}
\alias{create_sample}
\title{Create subsample of combined expression matrix}
\usage{
create_sample(
  combined_expr,
  batch_ids,
  sample_ids,
  sample_size = 1e+05,
  seed = 473
)
}
\description{
Create subsample of combined expression matrix
}
