% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_seu_batch_correct.R
\name{create_som_seurat}
\alias{create_som_seurat}
\title{Create Self-Organizing Map}
\usage{
create_som_seurat(
  object,
  markers = NULL,
  seed = 473,
  rlen = 10,
  mode = c("online", "batch", "pbatch"),
  cluster_method = c("kohonen", "lauvain", "leiden"),
  resolution = 0.8,
  xdim = 8,
  ydim = 8
)
}
\arguments{
\item{object}{A Seurat object}

\item{markers}{A vector of marker genes to use for the SOM. Defaults to all genes if NULL.}

\item{seed}{The seed to use when creating the SOM. Defaults to 473.}

\item{rlen}{Number of times the data is presented to the SOM network. Defaults to 10.}

\item{mode}{type of learning algorithm.}

\item{xdim}{The x-dimension size of the SOM. Defaults to 8.}

\item{ydim}{The y-dimension size of the SOM. Defaults to 8.}
}
\value{
A vector of clustering labels
}
\description{
The function uses the kohonen package to create a Self-Organizing Map (SOM).
It is used to segregate the cells for batch correction to make the correction less affected
by samples with high abundances of a particular cell type.
}
