% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{density_plots}
\alias{density_plots}
\title{Density ridges for two sets}
\usage{
density_plots(
  uncorrected,
  corrected,
  batch_ids,
  uncorrected_order = as.factor(batch_ids),
  corrected_order = as.factor(batch_ids),
  filename
)
}
\description{
Density ridges for two sets
}
