% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_batch_effect.R
\name{detect_batch_effect_express}
\alias{detect_batch_effect_express}
\title{Quicker function for detection of batch effects}
\usage{
detect_batch_effect_express(df, out_dir, downsample = NULL, seed = 472)
}
\arguments{
\item{df}{Tibble containing the expression data and batch information. See prepare_data.}

\item{out_dir}{Directory for plot output}

\item{downsample}{Number of cells to include in detection. If not specified all cells will be used.}
}
\description{
Quicker function for detection of batch effects
}
\examples{
detect_batch_effect_express(df = exprs, out_dir = '/my/cycombine/dir/')
detect_batch_effect_express(df = exprs, downsample = 100000, out_dir = '/my/cycombine/dir/')
}
