% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_export_data.R
\name{df2SCE}
\alias{df2SCE}
\title{Exporting a dataframe to SingleCellObject}
\usage{
df2SCE(
  df,
  markers = NULL,
  non_markers = NULL,
  sample_col = "sample",
  panel = NULL,
  panel_channel = "Channel",
  panel_antigen = "Marker",
  panel_type = "Type"
)
}
\arguments{
\item{df}{Tibble with expression values and metadata}

\item{markers}{Markers to include in exprs object of SCE. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{non_markers}{Non-markers to include as colData in SCE. If NULL, non_markers will be based on cyCombine::non_markers.}

\item{sample_col}{The name of the column containing sample IDs}

\item{panel}{Optional: Panel as a data.frame. Should have colnames Channel, Marker, Type unless otherwise specified in the panel_ args.}

\item{panel_channel}{Optional: Only used if panel is given. It is the column name in the panel data that contains the channel names}

\item{panel_antigen}{Optional: Only used if panel is given. It is the column name in the panel data that contains the antigen names}

\item{panel_type}{Optional: Only used if panel is given. It is the column name in the panel data that contains the antigen types (none, state, type).
"none" will be excluded from SCE.}
}
\description{
Conversion of dataframe into separate data and metadata objects for subsequent transformation.
Rename variable names to fit the requirements of SCE-based tools
}
\examples{
\dontrun{
sce <- df \%>\%
  df2SCE(markers = markers, non_markers = NULL, panel = panel)
  }
}
\concept{export}
