% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_export_data.R
\name{df2Seurat}
\alias{df2Seurat}
\title{Exporting a dataframe to Seurat Object}
\usage{
df2Seurat(
  df,
  markers = NULL,
  non_markers = NULL,
  scatter = NULL,
  clean_names = FALSE,
  sample_col = "sample",
  panel = NULL,
  panel_channel = "Channel",
  panel_antigen = "Marker",
  panel_type = NULL,
  transform_cofactor = 5,
  assay_name = "RNA"
)
}
\arguments{
\item{df}{Tibble with expression values and metadata}

\item{markers}{Markers to include in data layer of Seurat object. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{non_markers}{Non-markers to include as metadata in Seurat object. If NULL, non_markers will be based on cyCombine::non_markers.}

\item{scatter}{The scatter columns to add to the Seurat object.}

\item{clean_names}{Cleans marker names.}

\item{sample_col}{It is the column name in the df that contains the sample names. Defaults to 'sample'.}

\item{panel}{Optional: Panel as a data.frame. Should have colnames Channel, Marker, Type unless otherwise specified in the panel_ args.}

\item{panel_channel}{Optional: Only used if panel is given. It is the column name in the panel data that contains the channel names}

\item{panel_antigen}{Optional: Only used if panel is given. It is the column name in the panel data that contains the antigen names}

\item{panel_type}{Optional: Only used if panel is given. It is the column name in the panel data that contains the antigen types (none, state, type).
"none" will be excluded from Seurat object. Set to NULL to disregard.}

\item{transform_cofactor}{The cofactor to use when reverse-transforming to raw counts}

\item{assay_name}{Name for the assay in the Seurat object. Defaults to 'RNA'.}
}
\description{
Conversion of dataframe into Seurat object for downstream analysis.
Rename variable names to fit the requirements of Seurat-based tools
}
\examples{
\dontrun{
seurat_obj <- df \%>\%
  df2Seurat(markers = markers, non_markers = NULL, panel = panel)
  }
}
\seealso{
Other export: 
\code{\link{df2SCE}()},
\code{\link{sce2FCS}()}
}
\concept{export}
