% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_emd}
\alias{evaluate_emd}
\title{Evaluate EMD}
\usage{
evaluate_emd(
  preprocessed,
  corrected,
  cell_col = "label",
  batch_col = "batch",
  markers = NULL
)
}
\arguments{
\item{preprocessed}{Dataframe of uncorrected data}

\item{corrected}{Dataframe of corrected data}

\item{cell_col}{Column name of df that contains cell population labels (or clusters)}

\item{batch_col}{Column name of df that contains batch numbers}

\item{markers}{Vector of the markers to calculate EMD for. If NULL, \code{\link{get_markers}} will be used to find markers}
}
\description{
The function computes the Eartch Movers Distance of the two given datasets.
Then the reduction is calculated for every celltype
}
