% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_emd2}
\alias{evaluate_emd2}
\title{Evaluate EMD without population}
\usage{
evaluate_emd2(preprocessed, corrected, batch_col = "batch")
}
\description{
Evaluate EMD without population
}
