% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_mad}
\alias{evaluate_mad}
\title{Evaluate MAD}
\usage{
evaluate_mad(uncorrected, corrected, batch_col = "batch", markers = NULL)
}
\arguments{
\item{uncorrected}{Dataframe of uncorrected data}

\item{corrected}{Dataframe of corrected data}

\item{batch_col}{Column name of df that contains batch numbers}

\item{markers}{Vector of the markers to calculate EMD for. If NULL, \code{\link{get_markers}} will be used to find markers}
}
\description{
The function computes the Median Absolute Deviation of the two given datasets.
Then the reduction is calculated as the relative change in total MAD.
}
\seealso{
Other mad: 
\code{\link{compute_mad}()}
}
\concept{mad}
