% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{fcs_sample}
\alias{fcs_sample}
\title{Extract from a flowset given a sample of indices}
\usage{
fcs_sample(flowframe, sample, nrows, seed = 473)
}
\description{
Extract from a flowset given a sample of indices
}
