% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_panel_merging.R
\name{impute_across_panels}
\alias{impute_across_panels}
\title{Impute non-overlapping channels for whole data sets}
\usage{
impute_across_panels(
  impute_for,
  complete_obs,
  overlap_channels,
  impute_channels
)
}
\arguments{
\item{impute_for}{Dataframe with expression values to impute markers for}

\item{complete_obs}{Dataframe with expression values to base imputation on}

\item{overlap_channels}{Channels (markers) that overlap between impute_for and complete_obs, which can be used to base imputation on}

\item{impute_channels}{Channels to impute and add to the impute_for set (must be present in complete_obs)}
}
\description{
This function imputes the expression values for a whole dataset based on the overlapping markers contained in another dataset
The purpose is to be able to merge multi-panel data or to merge differently run datasets which have non-overlapping markers
}
\examples{
df_imputed <- df_panel1 \%>\%
  impute_across_panels(complete_obs = df_panel2, overlap_channels = intersect(df_panel1_markers, df_panel2_markers), impute_channels = unique_df_panel2_markers)
}
\seealso{
Other merging: 
\code{\link{salvage_problematic}()}
}
\concept{merging}
