% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{linear_shift}
\alias{linear_shift}
\title{Linearly shift data to lowest value to zero and all other values are shifted linearly along with it}
\usage{
linear_shift(df, markers = NULL, .keep = FALSE)
}
\arguments{
\item{df}{The dataframe to transform}

\item{markers}{The markers to transform on}

\item{.keep}{Keep all channels. If FALSE, channels that are not transformed are removed}
}
\description{
Linearly shift data to lowest value to zero and all other values are shifted linearly along with it
}
\examples{
preprocessed <- df \%>\%
  linear_shift(markers = markers)
}
\seealso{
Other preprocess: 
\code{\link{transform_asinh}()}
}
\concept{preprocess}
