% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_markers}
\alias{non_markers}
\title{Column names that are not cytometry markers}
\format{
An object of class \code{character} of length 8.
}
\usage{
non_markers
}
\description{
These are used to identify which markers not to manipulate by cyCombine functions.
The list can be manually expanded with 'non_markers <- c(non_markers, "new nonmarker")'
}
\keyword{datasets}
