% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{normalize}
\alias{normalize}
\title{Batch-wise normalization of data}
\usage{
normalize(
  df,
  markers = NULL,
  norm_method = c("scale", "rank", "qnorm"),
  ties.method = c("average", "first", "last", "random", "max", "min"),
  mc.cores = 1
)
}
\arguments{
\item{df}{tibble with expression values}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{norm_method}{Normalization method. Should be either 'rank', 'scale' or 'qnorm'. Default: 'scale'}

\item{ties.method}{The method to handle ties, when using rank. Default: 'average'. See ?rank for other options.}

\item{mc.cores}{Number of cores for parallelization}
}
\description{
This function normalizes the data in a batch-wise manner.
The purpose is to minimize the impact of batch correction when clustering the data prior to batch correction.
Three normalisation methods are implemented: Z-score, Rank, and Quantile normalization.
Z-score is recommended in cases where batches from a single study/experiment is merged.
Rank is recommend in cases where data from different studies/experiments are merged.
Quantile is not recommended.
}
\examples{
\dontrun{
df_normed <- df \%>\%
  normalize()
  }
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{batch_correct_seurat}()},
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{quantile_norm}()}
}
\concept{batch}
