% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_sce_batch_correct.R
\name{normalize_sce}
\alias{normalize_sce}
\title{Batch-wise Normalization of Data Using Bioconductor}
\usage{
normalize_sce(
  sce,
  markers = NULL,
  assay = "exprs",
  norm_assay = "normalized",
  norm_method = c("scale", "rank", "CLR", "lognorm", "none"),
  ties.method = c("average", "first", "last", "random", "max", "min"),
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{markers}{A character vector specifying the markers (features) to use for
normalization. If NULL, all features are used.}

\item{assay}{Character string specifying the assay in \code{sce} to use
as input (e.g., "counts", "logcounts"). Defaults to "counts".}

\item{norm_assay}{Character string for the name of the new assay where
normalized data will be stored. Defaults to "normdata"..}

\item{norm_method}{Normalization method: "scale" (Z-score), "rank", "clr",
"lognorm", or "qnorm" (Quantile normalization). Defaults to "scale".}

\item{ties.method}{The method to handle ties, when using rank. Default: 'average'. See ?rank for other options.}

\item{...}{For compatibility}
}
\value{
A SingleCellExperiment object with a new assay containing the
normalized data (named according to \code{new_assay_name}).
}
\description{
This function normalizes the data in a batch-wise manner using functions
compatible with SingleCellExperiment objects. The purpose is to minimize
the impact of batch effects when clustering the data prior to batch correction.
Four normalization methods are implemented: Z-score ("scale"), Rank,
CLR ("clr"), and Log-normalization ("lognorm").
}
\details{
Scale is recommended in most cases.
Rank sometimes performes better than scale.
CLR is recommended for ADT data (applied per feature across cells).
Log-normalization is a standard method for RNA-seq.
}
