% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_seu_batch_correct.R
\name{normalize_seurat}
\alias{normalize_seurat}
\title{Batch-wise Normalization of Data Using Seurat}
\usage{
normalize_seurat(
  object,
  markers = NULL,
  layer = "counts",
  assay = NULL,
  norm_method = c("scale", "rank", "CLR", "lognorm", "qnorm", "none"),
  ties.method = c("average", "first", "last", "random", "max", "min"),
  mc.cores = 1,
  pb = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{layer}{Layer to use from the Seurat object}

\item{norm_method}{Normalization method: "scale" (Z-score), "CLR", "lognorm", "rank", or "qnorm" (Quantile normalization). Defaults to "scale".}

\item{ties.method}{The method to handle ties, when using rank. Default: 'average'. See ?rank for other options.}

\item{mc.cores}{Number of cores for parallelization}

\item{pb}{Progress bar for parallelization}
}
\value{
A Seurat object with normalized data.
}
\description{
This function normalizes the data in a batch-wise manner using Seurat.
The purpose is to minimize the impact of batch effects when clustering the data prior to batch correction.
Four normalization methods are implemented: Z-score, Rank, CLR, and Quantile normalization.
Z-score is recommended for batches from a single study/experiment.
Rank is recommended for data from different studies/experiments.
CLR is recommended for ADT data.
Quantile is not recommended.
}
