% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_density}
\alias{plot_density}
\title{Density ridges for two sets}
\usage{
plot_density(
  uncorrected,
  corrected,
  markers = NULL,
  filename = NULL,
  y = "batch",
  xlim = 10
)
}
\description{
Density ridges for two sets
}
