% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Dimensionality reduction plot}
\usage{
plot_dimred(
  df,
  name,
  type = "pca",
  plot = "batch",
  markers = NULL,
  seed = 473,
  return_coord = F
)
}
\description{
Dimensionality reduction plot
}
