% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_dimred_full}
\alias{plot_dimred_full}
\title{Dimensionality reduction plots - colored with labels, batches and marker expression}
\usage{
plot_dimred_full(
  df,
  name,
  type = "umap",
  markers = NULL,
  seed = 473,
  out_dir = NULL
)
}
\arguments{
\item{df}{tibble of data to plot}

\item{name}{Name of the output plot}

\item{type}{Type of dimensionality reduction ("umap" or "pca")}

\item{markers}{Markers to include in dimensionality reduction}

\item{seed}{For reproducibility}

\item{out_dir}{Directory to put output figures}
}
\description{
Dimensionality reduction plots - colored with labels, batches and marker expression
}
