% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_save_four}
\alias{plot_save_four}
\title{Save four plots aligned with cowplot}
\usage{
plot_save_four(plot1, plot2, plot3, plot4, filename)
}
\description{
Save four plots aligned with cowplot
}
