% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_save_two}
\alias{plot_save_two}
\title{Save two plots aligned with cowplot}
\usage{
plot_save_two(plot1, plot2, filename)
}
\arguments{
\item{plot1}{Left plot in output}

\item{plot2}{Right plot in output}

\item{filename}{Filename of output}
}
\description{
Save two plots aligned with cowplot
}
