% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_umap_labels}
\alias{plot_umap_labels}
\title{Dimensionality reduction plots for a two-batch dataset before and after correction - colored by batch and label}
\usage{
plot_umap_labels(uncorrected, corrected, filename = NULL)
}
\description{
Dimensionality reduction plots for a two-batch dataset before and after correction - colored by batch and label
}
