% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{prepare_flowset}
\alias{prepare_flowset}
\title{Prepare and downsample flowset}
\usage{
prepare_flowset(
  flowset,
  batch_ids,
  sample_ids,
  sample_size = 1e+05,
  seed = 473
)
}
\description{
Prepare and downsample flowset
}
