% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess a directory of .fcs files}
\usage{
preprocess(
  data_dir,
  meta_file,
  markers,
  down_sample = TRUE,
  sample_size = 1e+05,
  seed = 473,
  cofactor = 5
)
}
\description{
Preprocess a directory of .fcs files
}
