% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_export_data.R
\name{process_fcs_files}
\alias{process_fcs_files}
\title{Process FCS files for FlowJo integration}
\usage{
process_fcs_files(input_folder, output_folder)
}
\arguments{
\item{input_folder}{path to input folder}

\item{output_folder}{path to output folder}
}
\description{
Process corrected and exported FCS files such that FlowJo can read them.
Reading FCS files in R with flowCore adds 'flowCore_' keywords that are
incompatible with FlowJo.
The \code{process_fcs_files} function reads a directory of corrected files
and exports them without the flowCore keywords.
}
\note{
flowCore write.FCS
}
\examples{
\dontrun{
 process_fcs_files(
 input_folder = "path/to/input_folder",
 output_folder = "path/to/output_folder")
  }

}
