% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{quantile_norm}
\alias{quantile_norm}
\title{Batch-wise quantile normalization per marker}
\usage{
quantile_norm(df, markers = NULL)
}
\arguments{
\item{df}{Dataframe with expression values}

\item{markers}{Markers to correct. If NULL, markers will be found using the \code{\link{get_markers}} function.}
}
\description{
This function quantile normalizes the data in a batch-wise manner.
The purpose is to minimize the impact of batch correction when clustering the data prior to batch correction.
}
\examples{
df_qnorm <- preprocessed \%>\%
  quantile_norm()
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{ranking}()},
\code{\link{scale_expr}()}
}
\concept{batch}
