% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{randomize_matrix}
\alias{randomize_matrix}
\title{Randomize a matrix of values}
\usage{
randomize_matrix(mat)
}
\arguments{
\item{mat}{A matrix to randomize}
}
\description{
Subtracts a random value between 0 and 1 from all values. Then convert negative values to zeros.
}
