% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run analysis of a batch correction}
\usage{
run_analysis(
  tool,
  data,
  data_dir,
  uncorrected_extension = "_uncorrected",
  corrected_extension = "_corrected",
  variant = NULL,
  uncorrected_variant = NULL,
  use_cycombine_uncor = FALSE,
  restart = FALSE,
  md = NULL,
  panel = NULL,
  markers = NULL,
  celltype_col = "label",
  segment = "",
  binSize = 0.1,
  rlen = 10,
  gridsize = 8,
  seed = 473,
  umap_size = 20000
)
}
\arguments{
\item{tool}{The name of the tool used to batch correct}

\item{data}{The name of the data used}

\item{data_dir}{The location of the uncorrected and corrected data}

\item{uncorrected_extension}{The extension used to name the uncorrected data. Default: "_uncorrected}

\item{corrected_extension}{The extension used to name the corrected data. Default: "_corrected"}

\item{variant}{Optional: A parameter to set a variant name of an experiment}

\item{use_cycombine_uncor}{If TRUE, the uncorrected data made by cyCombine will be used.}

\item{restart}{If TRUE, the SOM grid will be calculated even if it has been computed and stored previously}

\item{md}{Optional: Metadata filename. Currently not useful}

\item{panel}{Optional: If given, it will be used to define markers. Otherwise the function \code{\link{get_markers}} will be used}

\item{markers}{Optional: Manually define markers to use in plots and performance metrics}

\item{celltype_col}{Optional: If the cell types are known, specify which column they are defined in. If NULL, a clustering will be run.}

\item{segment}{Optional: Run only a specific segment of the analysis. Options include: "emd", "density", "umap"}

\item{rlen}{Number of times the data is presented to the SOM network}

\item{gridsize}{The gridsize to use when clustering. Only used if no celltype_col is given}

\item{seed}{The seed to use when creating the UMAP}
}
\description{
This function only runs under the assumption that the batch correction has been run with the cyCombine workflow in mind.
Some result preparation can be necessary, if cyCombine was not used during batch correction.
This function assumes that the uncorrected data us stored under the name \code{"{data_dir}/{tool}_{data}{variant}{uncorrected_extension}.RDS"} and
the corrected data is stored with the name \code{"{data_dir}/{tool}_{data}{variant}{corrected_extension}.RDS"}.
The analysis currently encompass marker-wise density plots, UMAP of corrected vs uncorrected, and Earth Movers Distance.
}
\examples{
run_analysis(tool = "cycombine",
 data = "dfci1",
 data_dir = "_data")
 run_analysis(tool = "cycombine",
 data = "FR-FCM-ZY34",
 data_dir = "_data",
 variant = "_p3",
 panel = "/attachments/MC_panel3.xlsx")
}
