% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_panel_merging.R
\name{salvage_problematic}
\alias{salvage_problematic}
\title{Salvaging problematic channels - one at a time}
\usage{
salvage_problematic(
  df,
  correct_batches,
  channel,
  label = NULL,
  sample_size = NULL,
  exclude = NULL,
  xdim = 8,
  ydim = 8,
  seed = 482
)
}
\arguments{
\item{df}{Dataframe with expression values}

\item{correct_batches}{Batches with misstaining (or other problem) for which to perform the imputation of 'channel'}

\item{channel}{Channel (marker) to impute}

\item{label}{(Optional) Column with group labels to base imputations on}

\item{sample_size}{Number of cells to base imputation on - defaults to all cells in df not in 'correct_batches'}

\item{exclude}{Channels to exclude (e.g. other misstained channels not to be used for imputation)}

\item{xdim}{The x-dimension size of the SOM.}

\item{ydim}{The y-dimension size of the SOM.}

\item{seed}{The seed to use when creating the SOM.}
}
\description{
This function imputes the expression values for a single channel based on the information in the other channels.
The purpose is to be able to salvage a misstained channel in single batches of an experiment
}
\examples{
\dontrun{
df_fixed <- df \%>\%
  salvage_problematic(correct_batches = 1, channel = 'CD127')
  }
}
\seealso{
Other merging: 
\code{\link{impute_across_panels}()}
}
\concept{merging}
