% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{save_four_plots}
\alias{save_four_plots}
\title{Save four plots aligned with cowplot}
\usage{
save_four_plots(plot1, plot2, plot3, plot4, filename)
}
\description{
Save four plots aligned with cowplot
}
