% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_export_data.R
\name{sce2FCS}
\alias{sce2FCS}
\title{Convert SingelCellExperiment into flowSet and store as FCS files}
\usage{
sce2FCS(
  sce,
  outdir = NULL,
  split_by = "sample_id",
  assay = "counts",
  keep_dr = TRUE,
  keep_cd = TRUE,
  randomize = FALSE
)
}
\arguments{
\item{sce}{SingleCellExperiment to write to FCS files}

\item{outdir}{If given, the flowSet will be stored in FCS files}

\item{split_by}{NULL or a character string 
specifying a \code{colData(x)} column to split by.}

\item{assay}{a character string specifying 
which assay data to use; valid values are \code{assayNames(x)}. 
When writing out FCS files, this should correspond to count-like data!}

\item{keep_cd, keep_dr}{logials specifying whether cell metadata 
(stored in \code{colData(x)}) and dimension reductions 
(stored in \code{reducedDims(x)}), respectively,
should be kept or dropped.}

\item{randomize}{(Default: FALSE) Logical determining whether counts are randomized for plotting purposes or not. Only works when assay = "counts" (default).}
}
\description{
Wrapper function that makes it easier to go from a SCE to flowSet and written FCS files.
The function uses CATALYST::sce2fcs and flowCore::write.flowSet to store the FCS files.
}
\examples{
\dontrun{
 sce2FCS(sce, outdir = "fcs_files")
  }
}
\seealso{
Other export: 
\code{\link{df2SCE}()},
\code{\link{df2Seurat}()}
}
\concept{export}
