% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correction.R
\name{batch_correct}
\alias{batch_correct}
\title{Run batch correction on preprocessed data}
\usage{
batch_correct(preprocessed_data, markers)
}
\description{
Run batch correction on preprocessed data
}
\seealso{
Other batch: 
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{scale_expr}()}
}
\concept{batch}
