% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{batch_correct}
\alias{batch_correct}
\title{Run batch correction on preprocessed data}
\usage{
batch_correct(
  df,
  label = NULL,
  xdim = 8,
  ydim = 8,
  parametric = TRUE,
  seed = 473,
  covar = NULL,
  markers = NULL,
  norm_method = "rank",
  ties.method = "average",
  som_type = "fsom"
)
}
\arguments{
\item{df}{Dataframe with expression values}

\item{label}{The cluster or cell type label. Either as a column name or vector.}

\item{xdim}{The x-dimension size of the SOM}

\item{ydim}{The y-dimension size of the SOM}

\item{parametric}{Default: TRUE. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used.}

\item{seed}{The seed to use when creating the SOM}

\item{covar}{The covariate ComBat should use. Can be a vector or a column name in the input datafrome.
If NULL, no covar will be used}

\item{markers}{Markers to scale. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{norm_method}{Normalization method. Should be either 'rank', 'scale' or 'qnorm'. Default: 'rank'}

\item{ties.method}{The method to handle ties, when using rank. Default: "min". See ?rank for other options.}

\item{som_type}{Default: "fsom". Determine which clustering function to use. FlowSOM or Kohonen. "fsom" sets to FlowSOM and anything else is Kohonen.}

\item{preprocessed}{The preprocessed dataframe to run batch correction on}
}
\description{
This is a wrapper function for the cyCombine batch correction workflow.
To run the workflow manually, type "batch_correct" to see the source code of this wrapper and follow along.
}
\seealso{
Other batch: 
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{normalize}()},
\code{\link{quantile_norm}()},
\code{\link{ranking}()},
\code{\link{scale_expr}()}
}
\concept{batch}
