% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_cyCombine.R
\name{batch_correct}
\alias{batch_correct}
\title{Run cyCombine batch correction}
\usage{
batch_correct(
  object,
  markers = NULL,
  metadata = NULL,
  label = NULL,
  xdim = 8,
  ydim = 8,
  rlen = 10,
  mode = c("online", "batch", "pbatch"),
  cluster_method = c("kohonen", "flowsom", "fusesom", "leiden", "kmeans"),
  resolution = 0.8,
  distf = c("euclidean", "sumofsquares", "cosine", "manhattan", "chebyshev"),
  nClus = NULL,
  seed = 473,
  assay = "exprs",
  layer = "data",
  norm_method = c("scale", "rank", "clr", "lognorm", "none"),
  ties.method = c("average", "first", "last", "random", "max", "min"),
  method = c("ComBat", "ComBat_seq"),
  covar = NULL,
  anchor = NULL,
  ref.batch = NULL,
  parametric = TRUE,
  mc.cores = 1,
  pb = FALSE
)
}
\arguments{
\item{object}{An object of type "data.frame", "matrix", "SingleCellExperiment", or "Seurat".}

\item{markers}{A character vector specifying the markers (features) to use for
normalization, clustering, and correction. If NULL, all features are used.}

\item{metadata}{Metadata to use in correction. Required when correcting a matrix}

\item{label}{Character string for the cluster labels column. If this column
already exists in the object metadata, the normalization and clustering steps are
skipped, and correction proceeds using these existing labels. Defaults to "Labels".}

\item{xdim}{Integer or vector of integers, the x-dimension(s) of the SOM grid
(for "kohonen" and "flowsom"). If a vector, the workflow iterates.}

\item{ydim}{Integer or vector of integers, the y-dimension(s) of the SOM grid.
If a vector, the workflow iterates. Length should match \code{xdim} or be 1.}

\item{rlen}{Integer, the number of iterations for SOM training.}

\item{mode}{Character, SOM training mode ("online", "batch", "pbatch") for "kohonen".}

\item{cluster_method}{Clustering method: "kohonen", "flowsom", "fusesom", "leiden", or "kmeans.}

\item{resolution}{Numeric, resolution parameter for Louvain/Leiden clustering.}

\item{distf}{Character, distance function for SOM.}

\item{nClus}{Integer, target number of clusters for FlowSOM metaclustering.}

\item{seed}{Integer, random seed for reproducibility.}

\item{assay}{Name of SingleCellExperiment assay to use}

\item{layer}{Name of Seurat layer to use.}

\item{norm_method}{Normalization method ("scale", "rank", "clr", "lognorm", "qnorm", "none").
Applied before clustering unless \code{labels_colname} is provided. Defaults to "scale".}

\item{ties.method}{Method for handling ties in rank normalization.}

\item{method}{Batch correction method ("ComBat", "ComBat_seq").}

\item{covar}{Optional: Covariate column name or vector for correction.}

\item{anchor}{Optional: Anchor column name or vector for correction.}

\item{ref.batch}{Optional: Reference batch for ComBat.}

\item{parametric}{Logical, for ComBat's parametric adjustment.}

\item{mc.cores}{Number of cores for parallel processing. Defaults to 1.}

\item{pb}{Logical, whether to show progress bars. Defaults to FALSE.}
}
\value{
An object of the same type as the input
}
\description{
As of cyCombine version 0.3.0, this function will run the cyCombine workflow
on any object of type "data.frame", "SingleCellExperiment", and "Seurat".
}
\details{
The workflow proceeds as follows:
\enumerate{
\item \strong{Normalization:} Normalizes data batch-wise using the method
specified by \code{norm_method}.
\item \strong{Clustering:} Clusters cells based on the normalized data using the chosen \code{cluster_method}.
This step is skipped if \code{label} is provided and already exists in the object.
\item \strong{Batch Correction:} Applies ComBat or ComBat-seq batch correction
within each cluster defined in step 2.
}

The process can be iterated by providing vectors for \code{xdim} and \code{ydim},
performing clustering and correction iteratively. In subsequent iterations, the
previously corrected data is used as input for normalization/clustering.
}
