% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_seu_batch_correct.R
\name{batch_correct_seurat}
\alias{batch_correct_seurat}
\title{Run batch correction on data}
\usage{
batch_correct_seurat(
  object,
  xdim = 8,
  ydim = 8,
  nClus = NULL,
  rlen = 10,
  mode = c("online", "batch", "pbatch"),
  parametric = TRUE,
  method = c("ComBat", "ComBat_seq"),
  cluster_method = c("kohonen", "flowsom", "lauvain", "leiden"),
  resolution = 0.8,
  ref.batch = NULL,
  seed = 473,
  label = NULL,
  covar = NULL,
  anchor = NULL,
  markers = NULL,
  layer = "counts",
  norm_method = c("scale", "rank", "CLR", "lognorm", "qnorm", "none"),
  ties.method = "average",
  mc.cores = 1,
  pb = FALSE
)
}
\arguments{
\item{object}{An object of type "data.frame", "matrix", "SingleCellExperiment", or "Seurat".}

\item{xdim}{Integer or vector of integers, the x-dimension(s) of the SOM grid
(for "kohonen" and "flowsom"). If a vector, the workflow iterates.}

\item{ydim}{Integer or vector of integers, the y-dimension(s) of the SOM grid.
If a vector, the workflow iterates. Length should match \code{xdim} or be 1.}

\item{nClus}{Integer, target number of clusters for FlowSOM metaclustering.}

\item{rlen}{Integer, the number of iterations for SOM training.}

\item{mode}{Character, SOM training mode ("online", "batch", "pbatch") for "kohonen".}

\item{parametric}{Logical, for ComBat's parametric adjustment.}

\item{method}{Batch correction method ("ComBat", "ComBat_seq").}

\item{cluster_method}{Clustering method: "kohonen", "flowsom", "fusesom", "leiden", or "kmeans.}

\item{resolution}{Numeric, resolution parameter for Louvain/Leiden clustering.}

\item{ref.batch}{Optional: Reference batch for ComBat.}

\item{seed}{Integer, random seed for reproducibility.}

\item{label}{Character string for the cluster labels column. If this column
already exists in the object metadata, the normalization and clustering steps are
skipped, and correction proceeds using these existing labels. Defaults to "Labels".}

\item{covar}{Optional: Covariate column name or vector for correction.}

\item{anchor}{Optional: Anchor column name or vector for correction.}

\item{markers}{A character vector specifying the markers (features) to use for
normalization, clustering, and correction. If NULL, all features are used.}

\item{layer}{Name of Seurat layer to use.}

\item{norm_method}{Normalization method ("scale", "rank", "clr", "lognorm", "qnorm", "none").
Applied before clustering unless \code{labels_colname} is provided. Defaults to "scale".}

\item{ties.method}{Method for handling ties in rank normalization.}

\item{mc.cores}{Number of cores for parallel processing. Defaults to 1.}

\item{pb}{Logical, whether to show progress bars. Defaults to FALSE.}
}
\description{
This is a wrapper function for the cyCombine batch correction workflow.
To run the workflow manually, type "batch_correct" to see the source code of this wrapper and follow along or read the vignettes on the GitHub page \url{https://github.com/biosurf/cyCombine}.
}
\examples{
\dontrun{
seu <- batch_correct_seurat(
 seu,
 markers = markers,
 covar = "condition"
 )
  }
}
\seealso{
Other batch: 
\code{\link{batch_correct_df}()},
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{normalize}()},
\code{\link{quantile_norm}()}
}
\concept{batch}
