% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_seu_batch_correct.R
\name{batch_correct_seurat}
\alias{batch_correct_seurat}
\title{Run batch correction on data}
\usage{
batch_correct_seurat(
  object,
  xdim = 8,
  ydim = 8,
  rlen = 10,
  mode = c("online", "batch", "pbatch"),
  parametric = TRUE,
  method = c("ComBat", "ComBat_seq"),
  cluster_method = c("kohonen", "flowsom", "lauvain", "leiden"),
  resolution = 0.8,
  ref.batch = NULL,
  seed = 473,
  label = NULL,
  covar = NULL,
  anchor = NULL,
  markers = NULL,
  layer = "counts",
  norm_method = c("scale", "rank", "CLR", "lognorm", "qnorm", "none"),
  ties.method = "average",
  mc.cores = 1,
  pb = FALSE
)
}
\arguments{
\item{object}{A Seurat object}

\item{xdim}{The x-dimension size of the SOM.}

\item{ydim}{The y-dimension size of the SOM.}

\item{rlen}{Number of times the data is presented to the SOM network}

\item{mode}{type of learning algorithm.}

\item{parametric}{Default: TRUE. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used.}

\item{method}{Default: "ComBat". Choose "ComBat" for cytometry data and "ComBat_seq" for bulk RNAseq data.}

\item{cluster_method}{Cluster method to use. Defaults to kohonen}

\item{resolution}{Resolution parameter for lauvain/leiden}

\item{ref.batch}{Optional. A string of the batch that should be used as the reference for batch adjustments.}

\item{seed}{The seed to use when creating the SOM.}

\item{label}{The cluster or cell type label. Either as a column name or vector.}

\item{covar}{The covariate ComBat should use. Can be a vector or a column name in the input tibble.
If NULL, no covar will be used}

\item{anchor}{Experimental: A column or vector specifying which samples are replicates and which are not. If specified, this column will be used as a covariate in ComBat. Be aware that it may be confounded with the condition.}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{layer}{Layer to use from the Seurat object}

\item{norm_method}{Normalization method. Should be either 'rank', 'scale', 'CLR', 'CLR_seu', 'CLR_med', or 'qnorm'. Default: 'scale'. CLR is recommended for ADT normalization. CLR_seu mimicks the Seurat implementation of CLR. CLR_med uses the median instead of geometric mean.}

\item{ties.method}{The method to handle ties, when using rank. Default: 'average'. See ?rank for other options.}

\item{mc.cores}{Number of cores for parallelization}

\item{pb}{Progress bar for parallelization}
}
\description{
This is a wrapper function for the cyCombine batch correction workflow.
To run the workflow manually, type "batch_correct" to see the source code of this wrapper and follow along or read the vignettes on the GitHub page \url{https://github.com/biosurf/cyCombine}.
}
\examples{
\dontrun{
seu <- batch_correct_seurat(
 seu,
 markers = markers,
 covar = "condition"
 )
  }
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{normalize}()},
\code{\link{quantile_norm}()}
}
\concept{batch}
