% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{compile_fcs}
\alias{compile_fcs}
\title{Compile all .fcs files in a directory to a flowset}
\usage{
compile_fcs(data_dir, pattern = "\\\\.fcs")
}
\arguments{
\item{data_dir}{Directory containing the .fcs files}

\item{pattern}{The pattern to use to find the files in the folder}
}
\description{
Compile all .fcs files in a directory to a flowset
}
\examples{
fcs <- compile_fcs(data_dir = "_data/raw", metadata = "meta/metadata.csv", sample_ids = "sample", batch_ids = "batch", filename_col = "fcs_filenames")
}
