% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{compile_fcs}
\alias{compile_fcs}
\title{Compile all .fcs files in a directory to a flowset}
\usage{
compile_fcs(data_dir, pattern = "\\\\.fcs")
}
\arguments{
\item{data_dir}{Directory containing the .fcs files}

\item{pattern}{The pattern to use to find the files in the folder}
}
\description{
A simple function to compile a directory of FCS files into a flowSet object using flowCore::read.flowSet().
Use the pattern argument to select only a subset of FCS files.
}
\examples{
fcs <- compile_fcs(data_dir = "_data/raw", pattern = "\\\\.fcs")
}
\seealso{
Other dataprep: 
\code{\link{convert_flowset}()},
\code{\link{linear_shift}()},
\code{\link{prepare_data}()},
\code{\link{transform_asinh}()}
}
\concept{dataprep}
