% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{compile_fcs}
\alias{compile_fcs}
\title{Compile all .fcs files in a directory to a flowset}
\usage{
compile_fcs(
  data_dir,
  metadata = NULL,
  sample_col = NULL,
  batch_col = "Batch",
  filename_col = "FCS_name",
  condition_col = NULL,
  pattern = "\\\\.fcs"
)
}
\arguments{
\item{data_dir}{Directory containing the .fcs files}

\item{metadata}{Can be either a filename or data.frame of the metadata file If NULL, sample and batch ids are not predicted nor returned}

\item{sample_col}{The column in the metadata containing the sample ids. If NULL, sample ids will be the file names}

\item{batch_col}{The column in the metadata containing the batch ids}

\item{filename_col}{The column in the metadata containing the fcs filenames}

\item{condition_col}{Optional: The column in the metadata containing the condition. Will be used as the covariate in ComBat, but can be specified later.}

\item{pattern}{The pattern to use to find the files in the folder}
}
\description{
Compile all .fcs files in a directory to a flowset
}
\examples{
fcs <- compile_fcs(data_dir = "_data/raw", metadata = "meta/metadata.csv", sample_col = "sample", batch_col = "batch", filename_col = "fcs_filenames")
}
