% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{compile_fcs}
\alias{compile_fcs}
\title{Compile all .fcs files in a directory}
\usage{
compile_fcs(data_dir, down_sample = TRUE, sample_size = 1e+05, seed = 473)
}
\arguments{
\item{data_dir}{Directory containing the .fcs files}
}
\description{
Compile all .fcs files in a directory
}
