% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{convert_flowset}
\alias{convert_flowset}
\title{Convert a flowset into a workable dataframe}
\usage{
convert_flowset(
  flowset,
  sample_ids,
  batch_ids,
  down_sample = TRUE,
  sample_size = 1e+05,
  seed = 473
)
}
\arguments{
\item{flowset}{The flowset to convert}

\item{sample_ids}{Vector containing the sample ids}

\item{batch_ids}{Vector containing the batch ids}

\item{down_sample}{If TRUE, the output will be down-sampled to size sample_size}

\item{sample_size}{The size to down-sample to}

\item{seed}{The seed to use for down-sampling}
}
\description{
Convert a flowset into a workable dataframe
}
