% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{convert_flowset}
\alias{convert_flowset}
\title{Convert a flowset into a workable dataframe}
\usage{
convert_flowset(
  flowset,
  sample_ids,
  batch_ids,
  down_sample = TRUE,
  sample_size = 1e+05,
  seed = 473
)
}
\description{
Convert a flowset into a workable dataframe
}
