% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{convert_flowset}
\alias{convert_flowset}
\title{Convert a flowset into a workable dataframe}
\usage{
convert_flowset(
  flowset,
  sample_ids = NULL,
  batch_ids = NULL,
  down_sample = TRUE,
  sample_size = 5e+05,
  seed = 473,
  panel = NULL,
  panel_channel = "fcs_colname",
  panel_antigen = "antigen"
)
}
\arguments{
\item{flowset}{The flowset to convert}

\item{sample_ids}{Vector containing the sample ids}

\item{batch_ids}{Vector containing the batch ids}

\item{down_sample}{If TRUE, the output will be down-sampled to size sample_size}

\item{sample_size}{The size to down-sample to}

\item{seed}{The seed to use for down-sampling}

\item{panel}{Panel as a data.frame. Is used to predict colnames}

\item{panel_channel}{Only used if panel is given. It is the column name in the panel data.frame that contains the channel names}

\item{panel_antigen}{Only used if panel is given. It is the column name in the panel data.frame that contains the antigen names}
}
\description{
Convert a flowset into a workable dataframe
}
\examples{
df <- convert_flowset(flowset = fcs$fcs_raw, sample_ids = fcs$sample_ids, batch_ids = fcs$batch_ids, down_sample = FALSE)
}
