% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correction.R
\name{correct_data}
\alias{correct_data}
\title{Correct data using ComBat}
\usage{
correct_data(combined_expr, som_classes, markers)
}
\description{
Correct data using ComBat
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{create_som}()},
\code{\link{scale_expr}()}
}
\concept{batch}
