% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{correct_data}
\alias{correct_data}
\title{Correct data using ComBat}
\usage{
correct_data(
  df,
  label,
  covar = NULL,
  anchor = NULL,
  markers = NULL,
  parametric = TRUE
)
}
\arguments{
\item{df}{Dataframe with expression values}

\item{label}{The cluster or cell type label. Either as a column name or vector.}

\item{covar}{The covariate ComBat should use (usually the sample condition, but other covariates may be used). Can be a vector or a column name in the input datafrome.
If NULL, no covar will be used.}

\item{anchor}{Optional: A column or vector specifying which samples are replicates and which are not. If specified, this column will be used as a covariate in ComBat. Be aware that it may be confounded with the condition.}

\item{markers}{Markers to scale. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{parametric}{Default: TRUE. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used.}
}
\description{
This function computes the batch correction on the preprocessed data using the ComBat algorithm.
}
\examples{
corrected <- preprocesed \%>\%
  correct_data(label = som_$unit.classif, covar = "condition")
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{create_som}()},
\code{\link{quantile_norm}()},
\code{\link{ranking}()},
\code{\link{scale_expr}()}
}
\concept{batch}
