% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correction.R
\name{correct_data}
\alias{correct_data}
\title{Correct data using ComBat}
\usage{
correct_data(df, som_classes, covar = NULL, markers = NULL)
}
\arguments{
\item{df}{Dataframe with expression values}

\item{som_classes}{The classes as returned by the \code{\link{create_som}} function}

\item{covar}{The covariate ComBat uses. If NULL, MAKE MORE FLEXIBLE!}
}
\description{
This function computes the batch correction on the preprocessed data using the ComBat algorithm.
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{create_som}()},
\code{\link{scale_expr}()}
}
\concept{batch}
