% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correction.R
\name{correct_data}
\alias{correct_data}
\title{Correct data using ComBat}
\usage{
correct_data(df, som_classes, covar = NULL, markers = NULL)
}
\arguments{
\item{df}{Dataframe with expression values}

\item{som_classes}{The classes as returned by the \code{\link{create_som}} function}

\item{covar}{The covariate ComBat uses. Can be a vector or a column name in the input datafrome.
If NULL, it is predicted based on the sample same (whether the sample id starts with HD or not)}

\item{markers}{Markers to scale. If NULL, markers will be found using the \code{\link{get_markers}} function.}
}
\description{
This function computes the batch correction on the preprocessed data using the ComBat algorithm.
}
\examples{
corrected <- preprocesed \%>\%
  correct_data(som_classes = som_$unit.classif, covar = covar)
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{create_som}()},
\code{\link{scale_expr}()}
}
\concept{batch}
