% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{correct_data}
\alias{correct_data}
\title{Correct data using ComBat}
\usage{
correct_data(
  df,
  label,
  markers = NULL,
  method = c("ComBat", "ComBat_seq"),
  covar = NULL,
  anchor = NULL,
  ref.batch = NULL,
  parametric = TRUE
)
}
\arguments{
\item{df}{tibble with expression values}

\item{label}{The cluster or cell type label. Either as a column name or vector.}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{method}{Default: "ComBat". Choose "ComBat" for cytometry data and "ComBat_seq" for bulk RNAseq data.}

\item{covar}{The covariate ComBat should use. Can be a vector or a column name in the input tibble.
If NULL, no covar will be used}

\item{anchor}{Experimental: A column or vector specifying which samples are replicates and which are not. If specified, this column will be used as a covariate in ComBat. Be aware that it may be confounded with the condition.}

\item{ref.batch}{Optional. A string of the batch that should be used as the reference for batch adjustments.}

\item{parametric}{Default: TRUE. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used.}
}
\description{
Compute the batch correction on the data using the ComBat algorithm.
Define a covariate, either as a character vector or name of tibble column.
The covariate should preferable be the cell condition types but can be any column that infers heterogeneity in the data.
The function assumes that the batch information is in the "batch" column and the data contains a "sample" column with sample information.
}
\examples{
\dontrun{
corrected <- uncorrected \%>\%
  correct_data(label = labels, covar = "condition")
  }
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{batch_correct_seurat}()},
\code{\link{create_som}()},
\code{\link{normalize}()},
\code{\link{quantile_norm}()}
}
\concept{batch}
