% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_seu_batch_correct.R
\name{correct_data_seurat}
\alias{correct_data_seurat}
\title{Correct data using ComBat}
\usage{
correct_data_seurat(
  object,
  markers = NULL,
  mc.cores = 1,
  method = c("ComBat", "ComBat_seq"),
  covar = NULL,
  anchor = NULL,
  ref.batch = NULL,
  parametric = TRUE,
  return_seurat = TRUE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{markers}{A vector of marker genes to use for the correction. Defaults to all genes if NULL.}

\item{mc.cores}{Number of cores for parallelization}

\item{method}{The method for batch correction. Choose "ComBat" for cytometry data and "ComBat_seq" for bulk RNAseq data. Defaults to "ComBat".}

\item{covar}{The covariate ComBat should use. Can be a vector or a metadata column name in the input Seurat object. If NULL, no covar will be used.}

\item{anchor}{A column or vector specifying which samples are replicates and which are not. If specified, this column will be used as a covariate in ComBat. Be aware that it may be confounded with the condition.}

\item{ref.batch}{Optional. A string of the batch that should be used as the reference for batch adjustments.}

\item{parametric}{Logical. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used. Defaults to TRUE.}

\item{return_seurat}{Logical. Whether the matrix or a Seurat object should be returned.}
}
\value{
A Seurat object with corrected data.
}
\description{
Compute the batch correction on the data using the ComBat algorithm.
Define a covariate, either as a character vector or name of Seurat metadata column.
The covariate should preferably be the cell condition types but can be any column that infers heterogeneity in the data.
The function assumes that the batch information is in the "batch" column and the data contains a "sample" column with sample information.
}
