% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{create_som}
\alias{create_som}
\title{Create Self-Organizing Map}
\usage{
create_som(
  df,
  markers = NULL,
  som_type = "fsom",
  seed = 473,
  xdim = 8,
  ydim = 8
)
}
\arguments{
\item{df}{tibble with expression values}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{som_type}{The clustering method to use. Default: "fsom" - running using FlowSOM. Set to "kohonen" to run with using kohonen.}

\item{seed}{The seed to use when creating the SOM.}

\item{xdim}{The x-dimension size of the SOM.}

\item{ydim}{The y-dimension size of the SOM.}
}
\value{
A vector of clustering labels
}
\description{
The function uses the FlowSOM or kohonen package to create a Self-Organizing Map.
It is used to segregate the cells for the batch correction to make the correction less affected
by samples with high abundances of a particular cell type.
FlowSOM is recommended with Z-score normalization.
kohonen is recommended with rank normalization
}
\examples{
labels <- uncorrected \%>\%
  create_som()
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{correct_data}()},
\code{\link{normalize}()},
\code{\link{quantile_norm}()}
}
\concept{batch}
