% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correction.R
\name{create_som}
\alias{create_som}
\title{Create SOM}
\usage{
create_som(scaled_expr, seed = 548, xdim = 10, ydim = 10)
}
\description{
Create SOM
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{correct_data_prev}()},
\code{\link{correct_data}()},
\code{\link{scale_expr}()}
}
\concept{batch}
