% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{create_som}
\alias{create_som}
\title{Create Self-Organizing Map}
\usage{
create_som(
  df,
  markers = NULL,
  seed = 473,
  cluster_method = c("kohonen", "flowsom", "fusesom", "kmeans"),
  distf = c("euclidean", "sumofsquares", "cosine", "manhattan", "chebyshev"),
  rlen = 10,
  mode = c("online", "batch", "pbatch"),
  xdim = 8,
  ydim = 8,
  nClus = NULL
)
}
\arguments{
\item{df}{tibble or seurat with expression values}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{seed}{The seed to use when creating the SOM.}

\item{cluster_method}{Cluster method to use. Defaults to kohonen}

\item{distf}{Distance metric used for kohonen ("sumofsquares", "euclidean", "manhattan") or FlowSOM ("euclidean", "cosine", "manhattan", "chebyshev")}

\item{rlen}{Number of times the data is presented to the SOM network}

\item{mode}{type of learning algorithm.}

\item{xdim}{The x-dimension size of the SOM.}

\item{ydim}{The y-dimension size of the SOM.}

\item{nClus}{(Usable with FlowSOM and kmeans) Number of clusters to export}
}
\value{
A vector of clustering labels
}
\description{
The function uses the kohonen package to create a Self-Organizing Map.
It is used to segregate the cells for the batch correction to make the correction less affected
by samples with high abundances of a particular cell type.
}
\examples{
\dontrun{
labels <- create_som(uncorrected)
  }
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{batch_correct_seurat}()},
\code{\link{correct_data}()},
\code{\link{normalize}()},
\code{\link{quantile_norm}()}
}
\concept{batch}
