% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_seu_batch_correct.R
\name{create_som_seurat}
\alias{create_som_seurat}
\title{Create Self-Organizing Map}
\usage{
create_som_seurat(
  object,
  layer = SeuratObject::Layers(object)[length(SeuratObject::Layers(object))],
  markers = NULL,
  seed = 473,
  rlen = 10,
  mode = c("online", "batch", "pbatch"),
  cluster_method = c("kohonen", "flowsom", "lauvain", "leiden"),
  resolution = 0.8,
  distf = c("euclidean", "sumofsquares", "cosine", "manhattan", "chebyshev"),
  xdim = 8,
  ydim = 8,
  nClus = NULL
)
}
\arguments{
\item{object}{A Seurat object}

\item{layer}{Layer to use from the Seurat object}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{seed}{The seed to use when creating the SOM.}

\item{rlen}{the number of times the complete data set will be
    presented to the network.}

\item{mode}{type of learning algorithm.}

\item{cluster_method}{Cluster method to use. Defaults to kohonen}

\item{resolution}{Resolution parameter for lauvain/leiden}

\item{distf}{Distance metric used for kohonen ("sumofsquares", "euclidean", "manhattan") or FlowSOM ("euclidean", "cosine", "manhattan", "chebyshev")}

\item{xdim}{The x-dimension size of the SOM.}

\item{ydim}{The y-dimension size of the SOM.}

\item{nClus}{(Usable with FlowSOM and kmeans) Number of clusters to export}
}
\value{
A vector of clustering labels
}
\description{
The function uses the kohonen package to create a Self-Organizing Map (SOM).
It is used to segregate the cells for batch correction to make the correction less affected
by samples with high abundances of a particular cell type.
}
