% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_cyCombine.R
\name{cyCombine}
\alias{cyCombine}
\title{Run the cyCombine workflow}
\usage{
cyCombine(object, ...)
}
\arguments{
\item{object}{An object of type "data.frame", "SingleCellExperiment", or "Seurat".}
}
\description{
This is an alias for the \code{batch_correct} function.
See \code{?batch_correct} for details.
}
