% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_batch_effect.R
\name{detect_batch_effect}
\alias{detect_batch_effect}
\title{Full function for detection of batch effects using cluster proportions}
\usage{
detect_batch_effect(
  df,
  downsample = NULL,
  xdim = 8,
  ydim = 8,
  seed = 382,
  markers = NULL,
  batch_col = "batch",
  label_col = "label",
  out_dir
)
}
\arguments{
\item{df}{Tibble containing the expression data and batch information. See prepare_data.}

\item{downsample}{Number of cells to include in detection. If not specified all cells will be used. One should be careful with the downsampling here as too strong downsampling leads to spurious results.}

\item{xdim}{Grid size in x-axis for SOM}

\item{ydim}{Grid size in y-axis for SOM}

\item{seed}{Random seed for reproducibility}

\item{markers}{If only some markers should be used this parameter is used to define them. If not set, all markers are used.}

\item{batch_col}{Name of column containing batch information}

\item{label_col}{If existing labels should be used, this column must be present in the data}

\item{out_dir}{Directory for plot output}
}
\description{
Full function for detection of batch effects using cluster proportions
}
