% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_export_data.R
\name{df2SCE}
\alias{df2SCE}
\title{Exporting a dataframe to SingleCellObject}
\usage{
df2SCE(
  df,
  markers = NULL,
  non_markers = NULL,
  scatter = NULL,
  clean_names = FALSE,
  sample_col = "sample",
  panel = NULL,
  panel_channel = "Channel",
  panel_antigen = "Marker",
  panel_type = NULL,
  transform_cofactor = 5
)
}
\arguments{
\item{df}{Tibble with expression values and metadata}

\item{markers}{Markers to include in exprs and counts object of SCE. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{non_markers}{Non-markers to include as colData in SCE. If NULL, non_markers will be based on cyCombine::non_markers.}

\item{sample_col}{It is the column name in the df that contains the sample names. Defaults to 'sample'.}

\item{panel}{Optional: Panel as a data.frame. Should have colnames Channel, Marker, Type unless otherwise specified in the panel_ args. Should be included if you want to store FCS files}

\item{panel_channel}{Optional: Only used if panel is given. It is the column name in the panel data that contains the channel names}

\item{panel_antigen}{Optional: Only used if panel is given. It is the column name in the panel data that contains the antigen names}

\item{panel_type}{Optional: Only used if panel is given. It is the column name in the panel data that contains the antigen types (none, state, type).
"none" will be excluded from SCE. Set to NULL to disregard.}

\item{transform_cofactor}{The cofactor to use when reverse-transforming to raw counts}
}
\description{
Conversion of dataframe into separate data and metadata objects for subsequent transformation.
Rename variable names to fit the requirements of SCE-based tools
}
\examples{
\dontrun{
sce <- df \%>\%
  df2SCE(markers = markers, non_markers = NULL, panel = panel)
  }
}
\seealso{
Other export: 
\code{\link{sce2FCS}()}
}
\concept{export}
