% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_emd}
\alias{evaluate_emd}
\title{Evaluate EMD}
\usage{
evaluate_emd(
  preprocessed,
  corrected,
  cell_col = "label",
  batch_col = "batch",
  non_markers = c("batch", "sample", "covar", "som", "label", "id")
)
}
\description{
Evaluate EMD
}
