% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correct.R
\name{normalize}
\alias{normalize}
\title{Batch-wise scaling of data}
\usage{
normalize(df, markers = NULL, norm_method = "rank", ties.method = "min")
}
\arguments{
\item{df}{Dataframe with expression values}

\item{markers}{Markers to normalize. If NULL, markers will be found using the \code{\link{get_markers}} function.}

\item{norm_method}{Normalization method. Should be either 'rank', 'scale' or 'qnorm'. Default: 'rank'}

\item{ties.method}{The method to handle ties, when using rank. Default: "min". See ?rank for other options.}
}
\description{
This function scales the data in a batch-wise manner.
The purpose is to minimize the impact of batch correction when clustering the data prior to batch correction.
}
\examples{
df_normed <- df \%>\%
  normalize()
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{correct_data}()},
\code{\link{create_som}()},
\code{\link{quantile_norm}()},
\code{\link{ranking}()},
\code{\link{scale_expr}()}
}
\concept{batch}
