% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_seu_batch_correct.R
\name{normalize_seurat}
\alias{normalize_seurat}
\title{Batch-wise Normalization of Data Using Seurat}
\usage{
normalize_seurat(
  object,
  markers = NULL,
  layer = "counts",
  norm_method = "scale",
  ties.method = "average",
  mc.cores = parallel::detectCores() - 1
)
}
\arguments{
\item{object}{A Seurat object}

\item{markers}{A vector of marker genes to normalize. Defaults to all genes if NULL.}

\item{layer}{Layer to use from the Seurat object}

\item{norm_method}{Normalization method: "scale" (Z-score), "rank", or "qnorm" (Quantile normalization). Defaults to "scale".}

\item{ties.method}{Method for handling ties in rank normalization. Options are "average", "first", "last", "random", "max", or "min". Defaults to "average".}

\item{mc.cores}{Number of cores for parallelization}
}
\value{
A Seurat object with normalized data.
}
\description{
This function normalizes the data in a batch-wise manner using Seurat.
The purpose is to minimize the impact of batch effects when clustering the data prior to batch correction.
Three normalization methods are implemented: Z-score, Rank, and Quantile normalization.
Z-score is recommended for batches from a single study/experiment.
Rank is recommended for data from different studies/experiments.
Quantile is not recommended.
}
