% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_density}
\alias{plot_density}
\title{Density ridges for two sets}
\usage{
plot_density(
  uncorrected,
  corrected,
  markers = NULL,
  filename,
  y = "batch",
  xlim = 10,
  dataset_names = NULL
)
}
\description{
Density ridges for two sets
}
\examples{
plot_density(uncorrected, corrected, y = 'batch', filename = 'my/dir/batchcor_plot.pdf')
plot_density(imputed1, imputed2, y = 'Type', dataset_names = paste('Panel', 1:2), filename = 'my/dir/merging_plot.pdf')
}
