% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_dimred_full}
\alias{plot_dimred_full}
\title{Dimensionality reduction plots - colored with labels, batches and marker expression}
\usage{
plot_dimred_full(
  df,
  name,
  type = "pca",
  markers = NULL,
  seed = 473,
  out_dir = NULL
)
}
\description{
Dimensionality reduction plots - colored with labels, batches and marker expression
}
