% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_save_four}
\alias{plot_save_four}
\title{Save four plots aligned with cowplot}
\usage{
plot_save_four(plot1, plot2, plot3, plot4, filename)
}
\arguments{
\item{plot1}{Top left plot in output}

\item{plot2}{Top right plot in output}

\item{plot3}{Lower left plot in output}

\item{plot4}{Lower right plot in output}

\item{filename}{Filename of output}
}
\description{
Save four plots aligned with cowplot
}
\seealso{
Other plot: 
\code{\link{plot_density}()},
\code{\link{plot_dimred_full}()},
\code{\link{plot_dimred}()},
\code{\link{plot_save_two}()},
\code{\link{plot_umap_labels}()}
}
\concept{plot}
