% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_umap_labels}
\alias{plot_umap_labels}
\title{Dimensionality reduction plots for a two-batch dataset before and after correction - colored by batch and label}
\usage{
plot_umap_labels(uncorrected, corrected, filename = NULL, seed = 473)
}
\arguments{
\item{seed}{For reproducibility}
}
\description{
Dimensionality reduction plots for a two-batch dataset before and after correction - colored by batch and label
}
\seealso{
Other plot: 
\code{\link{plot_density}()},
\code{\link{plot_dimred_full}()},
\code{\link{plot_dimred}()},
\code{\link{plot_save_four}()},
\code{\link{plot_save_two}()}
}
\concept{plot}
