% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess a directory of .fcs files}
\usage{
preprocess(data_dir, markers, sample_size = 3e+05, seed = 473)
}
\description{
Preprocess a directory of .fcs files
}
