% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correction.R
\name{scale_expr}
\alias{scale_expr}
\title{Batch-wise scaling of data}
\usage{
scale_expr(df)
}
\arguments{
\item{df}{Data.frame with expression values}
}
\description{
Batch-wise scaling of data
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{correct_data_prev}()},
\code{\link{correct_data}()},
\code{\link{create_som}()}
}
\concept{batch}
