% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_batch_correction.R
\name{scale_expr}
\alias{scale_expr}
\title{Batch-wise scaling of data}
\usage{
scale_expr(df)
}
\arguments{
\item{df}{Dataframe with expression values}
}
\description{
This function scales the data in a batch-wise manner.
The purpose is to minimize the impact of batch correction when clustering the data prior to batch correction.
}
\seealso{
Other batch: 
\code{\link{batch_correct}()},
\code{\link{correct_data}()},
\code{\link{create_som}()}
}
\concept{batch}
