% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{transform_asinh}
\alias{transform_asinh}
\title{Transform data using asinh}
\usage{
transform_asinh(df, markers = NULL, cofactor = 5, .keep = FALSE)
}
\arguments{
\item{df}{The dataframe to transform}

\item{markers}{The markers to transform on}

\item{cofactor}{The cofactor to use when transforming}

\item{.keep}{Keep all channels. If FALSE, channels that are not transformed are removed}
}
\description{
Inverse sine transformation of a dataframe.
}
\examples{
preprocessed <- df \%>\%
  transform_asinh(markers = markers)
}
\concept{preprocess}
