% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocess.R
\name{transform_asinh}
\alias{transform_asinh}
\title{Transform data using asinh}
\usage{
transform_asinh(input, markers, cofactor = 5, panel_fcs)
}
\description{
Transform data using asinh
}
\concept{preprocess}
